backend/services/agent_core/requirements.txt
openai>=1.3.0
Jinja2
python-dotenv
mlflow
fastapi
uvicorn
PyYAML
pgvector>=0.2.5
psycopg[binary,pool]>=3.1.19
jsonschema>=4.22.0
requests
PyMuPDF
camelot-py
